+++ packages/adguard-api-mv3/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.2.0 - 2024-10-02
 
 ### Changed
 
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `adguardApi.onRequestBlocking` event data now contains `companyCategoryName` property for `$redirect` rules [#137].
+
+### Fixed
+
+- Logging of `adguardApi.onRequestBlocking` events which are falsely considered as blocked requests.
 
 [#137]: https://github.com/AdguardTeam/tsurlfilter/issues/137